+++ Block/Adminhtml/System/Config/Form/Date.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Improntus\RetailRocket\Block\Adminhtml\System\Config\Form;
+
+use Magento\Backend\Block\Template\Context;
+use Magento\Framework\Stdlib\DateTime;
+
+/**
+ * Class Date
+ *
+ * @version 1.0.5
+ * @author Improntus <http://www.improntus.com> - Ecommerce done right
+ * @copyright Copyright (c) 2020 Improntus
+ * @package Improntus\RetailRocket\Block\Adminhtml\System\Config\Form
+ */
+class Date extends \Magento\Config\Block\System\Config\Form\Field
+{
+    /**
+     * Date constructor.
+     * @param Context $context
+     * @param array $data
+     */
+    public function __construct(
+        Context $context,
+        array $data = []
+    ) {
+        parent::__construct($context, $data);
+    }
+
+    /**
+     * @param \Magento\Framework\Data\Form\Element\AbstractElement $element
+     * @return string
+     */
+    public function render(\Magento\Framework\Data\Form\Element\AbstractElement $element)
+    {
+        $element->setDateFormat(DateTime::DATE_INTERNAL_FORMAT);
+        $element->setTimeFormat('HH:mm:ss');
+        $element->setShowsTime(true);
+        return parent::render($element);
+    }
+}